Network.AWS.CodeStar

Description

AWS CodeStar

This is the API reference for AWS CodeStar. This reference provides descriptions of the operations and data types for the AWS CodeStar API along with usage examples.

You can use the AWS CodeStar API to work with:

Projects and their resources, by calling the following:

DeleteProject , which deletes a project in AWS CodeStar.

DescribeProject , which lists the attributes of a project.

ListProjects , which lists all AWS CodeStar projects associated with your AWS account.

ListResources , which lists the resources associated with an AWS CodeStar project.

UpdateProject , which updates the attributes of an AWS CodeStar project.

Teams and team members, by calling the following:

AssociateTeamMember , which adds an IAM user to the team for an AWS CodeStar project.

DisassociateTeamMember , which removes an IAM user from the team for an AWS CodeStar project.

ListTeamMembers , which lists all the IAM users in the team for an AWS CodeStar project, including their roles and attributes.

Users, by calling the following:

CreateUserProfile , which creates a user profile that contains data associated with the user across all AWS CodeStar projects.

DeleteUserProfile , which deletes all user profile information across all AWS CodeStar projects.

DescribeUserProfile , which describes the profile of a user.

ListUserProfiles , which lists all AWS CodeStar user profiles.

UpdateUserProfile , which updates the profile for an AWS CodeStar user.

Service Configuration

codeStar :: Service

API version 2017-04-19 of the Amazon CodeStar SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CodeStar.

TeamMemberAlreadyAssociatedException

_TeamMemberAlreadyAssociatedException :: AsError a => Getting (First ServiceError) a ServiceError

The team member is already associated with a role in this project.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The specified input is either not valid, or it could not be validated.

InvalidServiceRoleException

_InvalidServiceRoleException :: AsError a => Getting (First ServiceError) a ServiceError

The service role is not valid.

ProjectCreationFailedException

_ProjectCreationFailedException :: AsError a => Getting (First ServiceError) a ServiceError

The project creation request was valid, but a nonspecific exception or error occurred during project creation. The project could not be created in AWS CodeStar.

UserProfileAlreadyExistsException

_UserProfileAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

A user profile with that name already exists in this region for the AWS account. AWS CodeStar user profile names must be unique within a region for the AWS account.

ProjectNotFoundException

_ProjectNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified AWS CodeStar project was not found.

TeamMemberNotFoundException

_TeamMemberNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified team member was not found.

ProjectAlreadyExistsException

_ProjectAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

An AWS CodeStar project with the same ID already exists in this region for the AWS account. AWS CodeStar project IDs must be unique within a region for the AWS account.

ProjectConfigurationException

_ProjectConfigurationException :: AsError a => Getting (First ServiceError) a ServiceError

Project configuration information is required but not specified.

ConcurrentModificationException

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

Another modification is being made. That modification must complete before you can make your change.

InvalidNextTokenException

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The next token is not valid.

UserProfileNotFoundException

_UserProfileNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The user profile was not found.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A resource limit has been exceeded.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListProjects

module Network.AWS.CodeStar.ListProjects

ListTeamMembers

module Network.AWS.CodeStar.ListTeamMembers

DeleteProject

module Network.AWS.CodeStar.DeleteProject

UpdateProject

module Network.AWS.CodeStar.UpdateProject

DisassociateTeamMember

module Network.AWS.CodeStar.DisassociateTeamMember

DescribeProject

module Network.AWS.CodeStar.DescribeProject

ListUserProfiles

module Network.AWS.CodeStar.ListUserProfiles

ListResources

module Network.AWS.CodeStar.ListResources

AssociateTeamMember

module Network.AWS.CodeStar.AssociateTeamMember

UpdateTeamMember

module Network.AWS.CodeStar.UpdateTeamMember

DescribeUserProfile

module Network.AWS.CodeStar.DescribeUserProfile

DeleteUserProfile

module Network.AWS.CodeStar.DeleteUserProfile

UpdateUserProfile

module Network.AWS.CodeStar.UpdateUserProfile

CreateUserProfile

module Network.AWS.CodeStar.CreateUserProfile

CreateProject

module Network.AWS.CodeStar.CreateProject

Types

ProjectSummary

data ProjectSummary

Information about the metadata for a project.

See: projectSummary smart constructor.

projectSummary :: ProjectSummary

Creates a value of ProjectSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

psProjectARN - The Amazon Resource Name (ARN) of the project.

psProjectId - The ID of the project.

psProjectARN :: Lens' ProjectSummary (Maybe Text)

The Amazon Resource Name (ARN) of the project.

psProjectId :: Lens' ProjectSummary (Maybe Text)

The ID of the project.

Resource

data Resource

Information about a resource for a project.

See: resource smart constructor.

resource

Arguments

:: Text
rId

-> Resource

Creates a value of Resource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rId - The Amazon Resource Name (ARN) of the resource.

rId :: Lens' Resource Text

The Amazon Resource Name (ARN) of the resource.

TeamMember

data TeamMember

Information about a team member in a project.

See: teamMember smart constructor.

teamMember

Arguments

:: Text
tmUserARN

-> Text
tmProjectRole

-> TeamMember

Creates a value of TeamMember with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tmRemoteAccessAllowed - Whether the user is allowed to remotely access project resources using an SSH public/private key pair.

tmUserARN - The Amazon Resource Name (ARN) of the user in IAM.

tmProjectRole - The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

tmRemoteAccessAllowed :: Lens' TeamMember (Maybe Bool)

Whether the user is allowed to remotely access project resources using an SSH public/private key pair.

tmUserARN :: Lens' TeamMember Text

The Amazon Resource Name (ARN) of the user in IAM.

tmProjectRole :: Lens' TeamMember Text

The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

UserProfileSummary

data UserProfileSummary

Information about a user's profile in AWS CodeStar.

See: userProfileSummary smart constructor.

userProfileSummary :: UserProfileSummary

Creates a value of UserProfileSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upsSshPublicKey - The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

upsUserARN - The Amazon Resource Name (ARN) of the user in IAM.

upsEmailAddress - The email address associated with the user.

upsDisplayName - The display name of a user in AWS CodeStar. For example, this could be set to both first and last name ("Mary Major") or a single name (Mary). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space (MJ, not MM).

upsSshPublicKey :: Lens' UserProfileSummary (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

upsUserARN :: Lens' UserProfileSummary (Maybe Text)

The Amazon Resource Name (ARN) of the user in IAM.

upsEmailAddress :: Lens' UserProfileSummary (Maybe Text)

The email address associated with the user.

upsDisplayName :: Lens' UserProfileSummary (Maybe Text)

The display name of a user in AWS CodeStar. For example, this could be set to both first and last name ("Mary Major") or a single name (Mary). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space (MJ, not MM).