Network.AWS.CodeStar.AssociateTeamMember

Description

Adds an IAM user to the team for an AWS CodeStar project.

Creating a Request

associateTeamMember

Arguments

:: Text
atmProjectId

-> Text
atmUserARN

-> Text
atmProjectRole

-> AssociateTeamMember

Creates a value of AssociateTeamMember with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atmRemoteAccessAllowed - Whether the team member is allowed to use an SSH public/private key pair to remotely access project resources, for example Amazon EC2 instances.

atmClientRequestToken - A user- or system-generated token that identifies the entity that requested the team member association to the project. This token can be used to repeat the request.

atmProjectId - The ID of the project to which you will add the IAM user.

atmUserARN - The Amazon Resource Name (ARN) for the IAM user you want to add to the DevHub project.

atmProjectRole - The AWS CodeStar project role that will apply to this user. This role determines what actions a user can take in an AWS CodeStar project.

data AssociateTeamMember

See: associateTeamMember smart constructor.

Request Lenses

atmRemoteAccessAllowed :: Lens' AssociateTeamMember (Maybe Bool)

Whether the team member is allowed to use an SSH public/private key pair to remotely access project resources, for example Amazon EC2 instances.

atmClientRequestToken :: Lens' AssociateTeamMember (Maybe Text)

A user- or system-generated token that identifies the entity that requested the team member association to the project. This token can be used to repeat the request.

atmProjectId :: Lens' AssociateTeamMember Text

The ID of the project to which you will add the IAM user.

atmUserARN :: Lens' AssociateTeamMember Text

The Amazon Resource Name (ARN) for the IAM user you want to add to the DevHub project.

atmProjectRole :: Lens' AssociateTeamMember Text

The AWS CodeStar project role that will apply to this user. This role determines what actions a user can take in an AWS CodeStar project.

Destructuring the Response

associateTeamMemberResponse

Arguments

:: Int
atmrsResponseStatus

-> AssociateTeamMemberResponse

Creates a value of AssociateTeamMemberResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atmrsClientRequestToken - The user- or system-generated token from the initial request that can be used to repeat the request.

atmrsResponseStatus - -- | The response status code.

data AssociateTeamMemberResponse

See: associateTeamMemberResponse smart constructor.

Response Lenses

atmrsClientRequestToken :: Lens' AssociateTeamMemberResponse (Maybe Text)

The user- or system-generated token from the initial request that can be used to repeat the request.

atmrsResponseStatus :: Lens' AssociateTeamMemberResponse Int

- | The response status code.