Network.AWS.CodeStar.CreateProject

Description

Reserved for future use. To create a project, use the AWS CodeStar console.

Creating a Request

createProject

Arguments

:: Text
cpName

-> Text
cpId

-> CreateProject

Creates a value of CreateProject with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpClientRequestToken - Reserved for future use.

cpDescription - Reserved for future use.

cpName - Reserved for future use.

cpId - Reserved for future use.

data CreateProject

See: createProject smart constructor.

Request Lenses

cpClientRequestToken :: Lens' CreateProject (Maybe Text)

Reserved for future use.

cpDescription :: Lens' CreateProject (Maybe Text)

Reserved for future use.

cpName :: Lens' CreateProject Text

Reserved for future use.

cpId :: Lens' CreateProject Text

Reserved for future use.

Destructuring the Response

createProjectResponse

Arguments

:: Int
cprsResponseStatus

-> Text
cprsId

-> Text
cprsArn

-> CreateProjectResponse

Creates a value of CreateProjectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cprsProjectTemplateId - Reserved for future use.

cprsClientRequestToken - Reserved for future use.

cprsResponseStatus - -- | The response status code.

cprsId - Reserved for future use.

cprsArn - Reserved for future use.

data CreateProjectResponse

See: createProjectResponse smart constructor.

Response Lenses

cprsProjectTemplateId :: Lens' CreateProjectResponse (Maybe Text)

Reserved for future use.

cprsClientRequestToken :: Lens' CreateProjectResponse (Maybe Text)

Reserved for future use.

cprsResponseStatus :: Lens' CreateProjectResponse Int

- | The response status code.

cprsId :: Lens' CreateProjectResponse Text

Reserved for future use.

cprsArn :: Lens' CreateProjectResponse Text

Reserved for future use.