Network.AWS.CodeStar.CreateUserProfile

Description

Creates a profile for a user that includes user preferences, such as the display name and email address assocciated with the user, in AWS CodeStar. The user profile is not project-specific. Information in the user profile is displayed wherever the user's information appears to other users in AWS CodeStar.

Creating a Request

createUserProfile

Arguments

:: Text
cupUserARN

-> Text
cupDisplayName

-> Text
cupEmailAddress

-> CreateUserProfile

Creates a value of CreateUserProfile with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cupSshPublicKey - The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

cupUserARN - The Amazon Resource Name (ARN) of the user in IAM.

cupDisplayName - The name that will be displayed as the friendly name for the user in AWS CodeStar.

cupEmailAddress - The email address that will be displayed as part of the user's profile in AWS CodeStar.

data CreateUserProfile

See: createUserProfile smart constructor.

Request Lenses

cupSshPublicKey :: Lens' CreateUserProfile (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

cupUserARN :: Lens' CreateUserProfile Text

The Amazon Resource Name (ARN) of the user in IAM.

cupDisplayName :: Lens' CreateUserProfile Text

The name that will be displayed as the friendly name for the user in AWS CodeStar.

cupEmailAddress :: Lens' CreateUserProfile Text

The email address that will be displayed as part of the user's profile in AWS CodeStar.

Destructuring the Response

createUserProfileResponse

Arguments

:: Int
cuprsResponseStatus

-> Text
cuprsUserARN

-> CreateUserProfileResponse

Creates a value of CreateUserProfileResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cuprsLastModifiedTimestamp - The date the user profile was last modified, in timestamp format.

cuprsSshPublicKey - The SSH public key associated with the user in AWS CodeStar. This is the public portion of the public/private keypair the user can use to access project resources if a project owner allows the user remote access to those resources.

cuprsEmailAddress - The email address that is displayed as part of the user's profile in AWS CodeStar.

cuprsDisplayName - The name that is displayed as the friendly name for the user in AWS CodeStar.

cuprsCreatedTimestamp - The date the user profile was created, in timestamp format.

cuprsResponseStatus - -- | The response status code.

cuprsUserARN - The Amazon Resource Name (ARN) of the user in IAM.

data CreateUserProfileResponse

See: createUserProfileResponse smart constructor.

Response Lenses

cuprsLastModifiedTimestamp :: Lens' CreateUserProfileResponse (Maybe UTCTime)

The date the user profile was last modified, in timestamp format.

cuprsSshPublicKey :: Lens' CreateUserProfileResponse (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. This is the public portion of the public/private keypair the user can use to access project resources if a project owner allows the user remote access to those resources.

cuprsEmailAddress :: Lens' CreateUserProfileResponse (Maybe Text)

The email address that is displayed as part of the user's profile in AWS CodeStar.

cuprsDisplayName :: Lens' CreateUserProfileResponse (Maybe Text)

The name that is displayed as the friendly name for the user in AWS CodeStar.

cuprsCreatedTimestamp :: Lens' CreateUserProfileResponse (Maybe UTCTime)

The date the user profile was created, in timestamp format.

cuprsResponseStatus :: Lens' CreateUserProfileResponse Int

- | The response status code.

cuprsUserARN :: Lens' CreateUserProfileResponse Text

The Amazon Resource Name (ARN) of the user in IAM.