Network.AWS.CodeStar.DescribeProject

Description

Describes a project and its resources.

Creating a Request

describeProject

Arguments

:: Text
dId

-> DescribeProject

Creates a value of DescribeProject with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dId - The ID of the project.

data DescribeProject

See: describeProject smart constructor.

Request Lenses

dId :: Lens' DescribeProject Text

The ID of the project.

Destructuring the Response

describeProjectResponse

Arguments

:: Int
drsResponseStatus

-> DescribeProjectResponse

Creates a value of DescribeProjectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsArn - The Amazon Resource Name (ARN) for the project.

drsProjectTemplateId - The ID for the AWS CodeStar project template used to create the project.

drsName - The display name for the project.

drsId - The ID of the project.

drsStackId - The ID of the primary stack in AWS CloudFormation used to generate resources for the project.

drsClientRequestToken - A user- or system-generated token that identifies the entity that requested project creation.

drsCreatedTimeStamp - The date and time the project was created, in timestamp format.

drsDescription - The description of the project, if any.

drsResponseStatus - -- | The response status code.

data DescribeProjectResponse

See: describeProjectResponse smart constructor.

Response Lenses

drsArn :: Lens' DescribeProjectResponse (Maybe Text)

The Amazon Resource Name (ARN) for the project.

drsProjectTemplateId :: Lens' DescribeProjectResponse (Maybe Text)

The ID for the AWS CodeStar project template used to create the project.

drsName :: Lens' DescribeProjectResponse (Maybe Text)

The display name for the project.

drsId :: Lens' DescribeProjectResponse (Maybe Text)

The ID of the project.

drsStackId :: Lens' DescribeProjectResponse (Maybe Text)

The ID of the primary stack in AWS CloudFormation used to generate resources for the project.

drsClientRequestToken :: Lens' DescribeProjectResponse (Maybe Text)

A user- or system-generated token that identifies the entity that requested project creation.

drsCreatedTimeStamp :: Lens' DescribeProjectResponse (Maybe UTCTime)

The date and time the project was created, in timestamp format.

drsDescription :: Lens' DescribeProjectResponse (Maybe Text)

The description of the project, if any.

drsResponseStatus :: Lens' DescribeProjectResponse Int

- | The response status code.