Network.AWS.CodeStar.DescribeUserProfile

Description

Describes a user in AWS CodeStar and the user attributes across all projects.

Creating a Request

describeUserProfile

Arguments

:: Text
dupUserARN

-> DescribeUserProfile

Creates a value of DescribeUserProfile with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dupUserARN - The Amazon Resource Name (ARN) of the user.

data DescribeUserProfile

See: describeUserProfile smart constructor.

Request Lenses

dupUserARN :: Lens' DescribeUserProfile Text

The Amazon Resource Name (ARN) of the user.

Destructuring the Response

describeUserProfileResponse

Arguments

:: Int
duprsResponseStatus

-> Text
duprsUserARN

-> UTCTime
duprsCreatedTimestamp

-> UTCTime
duprsLastModifiedTimestamp

-> DescribeUserProfileResponse

Creates a value of DescribeUserProfileResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

duprsSshPublicKey - The SSH public key associated with the user. This SSH public key is associated with the user profile, and can be used in conjunction with the associated private key for access to project resources, such as Amazon EC2 instances, if a project owner grants remote access to those resources.

duprsEmailAddress - The email address for the user. Optional.

duprsDisplayName - The display name shown for the user in AWS CodeStar projects. For example, this could be set to both first and last name ("Mary Major") or a single name (Mary). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space (MJ, not MM).

duprsResponseStatus - -- | The response status code.

duprsUserARN - The Amazon Resource Name (ARN) of the user.

duprsCreatedTimestamp - The date and time when the user profile was created in AWS CodeStar, in timestamp format.

duprsLastModifiedTimestamp - The date and time when the user profile was last modified, in timestamp format.

data DescribeUserProfileResponse

See: describeUserProfileResponse smart constructor.

Response Lenses

duprsSshPublicKey :: Lens' DescribeUserProfileResponse (Maybe Text)

The SSH public key associated with the user. This SSH public key is associated with the user profile, and can be used in conjunction with the associated private key for access to project resources, such as Amazon EC2 instances, if a project owner grants remote access to those resources.

duprsEmailAddress :: Lens' DescribeUserProfileResponse (Maybe Text)

The email address for the user. Optional.

duprsDisplayName :: Lens' DescribeUserProfileResponse (Maybe Text)

The display name shown for the user in AWS CodeStar projects. For example, this could be set to both first and last name ("Mary Major") or a single name (Mary). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space (MJ, not MM).

duprsResponseStatus :: Lens' DescribeUserProfileResponse Int

- | The response status code.

duprsUserARN :: Lens' DescribeUserProfileResponse Text

The Amazon Resource Name (ARN) of the user.

duprsCreatedTimestamp :: Lens' DescribeUserProfileResponse UTCTime

The date and time when the user profile was created in AWS CodeStar, in timestamp format.

duprsLastModifiedTimestamp :: Lens' DescribeUserProfileResponse UTCTime

The date and time when the user profile was last modified, in timestamp format.