Network.AWS.CodeStar.ListProjects

Description

Lists all projects in AWS CodeStar associated with your AWS account.

Creating a Request

listProjects :: ListProjects

Creates a value of ListProjects with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpNextToken - The continuation token to be used to return the next set of results, if the results cannot be returned in one response.

lpMaxResults - The maximum amount of data that can be contained in a single set of results.

data ListProjects

See: listProjects smart constructor.

Request Lenses

lpNextToken :: Lens' ListProjects (Maybe Text)

The continuation token to be used to return the next set of results, if the results cannot be returned in one response.

lpMaxResults :: Lens' ListProjects (Maybe Natural)

The maximum amount of data that can be contained in a single set of results.

Destructuring the Response

listProjectsResponse

Arguments

:: Int
lprsResponseStatus

-> ListProjectsResponse

Creates a value of ListProjectsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lprsNextToken - The continuation token to use when requesting the next set of results, if there are more results to be returned.

lprsResponseStatus - -- | The response status code.

lprsProjects - A list of projects.

data ListProjectsResponse

See: listProjectsResponse smart constructor.

Response Lenses

lprsNextToken :: Lens' ListProjectsResponse (Maybe Text)

The continuation token to use when requesting the next set of results, if there are more results to be returned.

lprsResponseStatus :: Lens' ListProjectsResponse Int

- | The response status code.

lprsProjects :: Lens' ListProjectsResponse [ProjectSummary]

A list of projects.