Network.AWS.CodeStar.ListUserProfiles

Description

Lists all the user profiles configured for your AWS account in AWS CodeStar.

Creating a Request

listUserProfiles :: ListUserProfiles

Creates a value of ListUserProfiles with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lupNextToken - The continuation token for the next set of results, if the results cannot be returned in one response.

lupMaxResults - The maximum number of results to return in a response.

data ListUserProfiles

See: listUserProfiles smart constructor.

Request Lenses

lupNextToken :: Lens' ListUserProfiles (Maybe Text)

The continuation token for the next set of results, if the results cannot be returned in one response.

lupMaxResults :: Lens' ListUserProfiles (Maybe Natural)

The maximum number of results to return in a response.

Destructuring the Response

listUserProfilesResponse

Arguments

:: Int
luprsResponseStatus

-> ListUserProfilesResponse

Creates a value of ListUserProfilesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

luprsNextToken - The continuation token to use when requesting the next set of results, if there are more results to be returned.

luprsResponseStatus - -- | The response status code.

luprsUserProfiles - All the user profiles configured in AWS CodeStar for an AWS account.

data ListUserProfilesResponse

See: listUserProfilesResponse smart constructor.

Response Lenses

luprsNextToken :: Lens' ListUserProfilesResponse (Maybe Text)

The continuation token to use when requesting the next set of results, if there are more results to be returned.

luprsResponseStatus :: Lens' ListUserProfilesResponse Int

- | The response status code.

luprsUserProfiles :: Lens' ListUserProfilesResponse [UserProfileSummary]

All the user profiles configured in AWS CodeStar for an AWS account.