Network.AWS.CodeStar.Types

Description

Service Configuration

codeStar :: Service

API version 2017-04-19 of the Amazon CodeStar SDK configuration.

Errors

_TeamMemberAlreadyAssociatedException :: AsError a => Getting (First ServiceError) a ServiceError

The team member is already associated with a role in this project.

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The specified input is either not valid, or it could not be validated.

_InvalidServiceRoleException :: AsError a => Getting (First ServiceError) a ServiceError

The service role is not valid.

_ProjectCreationFailedException :: AsError a => Getting (First ServiceError) a ServiceError

The project creation request was valid, but a nonspecific exception or error occurred during project creation. The project could not be created in AWS CodeStar.

_UserProfileAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

A user profile with that name already exists in this region for the AWS account. AWS CodeStar user profile names must be unique within a region for the AWS account.

_ProjectNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified AWS CodeStar project was not found.

_TeamMemberNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified team member was not found.

_ProjectAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

An AWS CodeStar project with the same ID already exists in this region for the AWS account. AWS CodeStar project IDs must be unique within a region for the AWS account.

_ProjectConfigurationException :: AsError a => Getting (First ServiceError) a ServiceError

Project configuration information is required but not specified.

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

Another modification is being made. That modification must complete before you can make your change.

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The next token is not valid.

_UserProfileNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The user profile was not found.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A resource limit has been exceeded.

ProjectSummary

data ProjectSummary

Information about the metadata for a project.

See: projectSummary smart constructor.

projectSummary :: ProjectSummary

Creates a value of ProjectSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

psProjectARN - The Amazon Resource Name (ARN) of the project.

psProjectId - The ID of the project.

psProjectARN :: Lens' ProjectSummary (Maybe Text)

The Amazon Resource Name (ARN) of the project.

psProjectId :: Lens' ProjectSummary (Maybe Text)

The ID of the project.

Resource

data Resource

Information about a resource for a project.

See: resource smart constructor.

resource

Arguments

:: Text
rId

-> Resource

Creates a value of Resource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rId - The Amazon Resource Name (ARN) of the resource.

rId :: Lens' Resource Text

The Amazon Resource Name (ARN) of the resource.

TeamMember

data TeamMember

Information about a team member in a project.

See: teamMember smart constructor.

teamMember

Arguments

:: Text
tmUserARN

-> Text
tmProjectRole

-> TeamMember

Creates a value of TeamMember with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tmRemoteAccessAllowed - Whether the user is allowed to remotely access project resources using an SSH public/private key pair.

tmUserARN - The Amazon Resource Name (ARN) of the user in IAM.

tmProjectRole - The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

tmRemoteAccessAllowed :: Lens' TeamMember (Maybe Bool)

Whether the user is allowed to remotely access project resources using an SSH public/private key pair.

tmUserARN :: Lens' TeamMember Text

The Amazon Resource Name (ARN) of the user in IAM.

tmProjectRole :: Lens' TeamMember Text

The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

UserProfileSummary

data UserProfileSummary

Information about a user's profile in AWS CodeStar.

See: userProfileSummary smart constructor.

userProfileSummary :: UserProfileSummary

Creates a value of UserProfileSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upsSshPublicKey - The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

upsUserARN - The Amazon Resource Name (ARN) of the user in IAM.

upsEmailAddress - The email address associated with the user.

upsDisplayName - The display name of a user in AWS CodeStar. For example, this could be set to both first and last name ("Mary Major") or a single name (Mary). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space (MJ, not MM).

upsSshPublicKey :: Lens' UserProfileSummary (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

upsUserARN :: Lens' UserProfileSummary (Maybe Text)

The Amazon Resource Name (ARN) of the user in IAM.

upsEmailAddress :: Lens' UserProfileSummary (Maybe Text)

The email address associated with the user.

upsDisplayName :: Lens' UserProfileSummary (Maybe Text)

The display name of a user in AWS CodeStar. For example, this could be set to both first and last name ("Mary Major") or a single name (Mary). The display name is also used to generate the initial icon associated with the user in AWS CodeStar projects. If spaces are included in the display name, the first character that appears after the space will be used as the second character in the user initial icon. The initial icon displays a maximum of two characters, so a display name with more than one space (for example "Mary Jane Major") would generate an initial icon using the first character and the first character after the space (MJ, not MM).