Network.AWS.CodeStar.UpdateTeamMember

Description

Updates a team member's attributes in an AWS CodeStar project. For example, you can change a team member's role in the project, or change whether they have remote access to project resources.

Creating a Request

updateTeamMember

Arguments

:: Text
utmProjectId

-> Text
utmUserARN

-> UpdateTeamMember

Creates a value of UpdateTeamMember with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

utmRemoteAccessAllowed - Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile. Even if this is set to True, the user must associate a public key with their profile before the user can access resources.

utmProjectRole - The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

utmProjectId - The ID of the project.

utmUserARN - The Amazon Resource Name (ARN) of the user for whom you want to change team membership attributes.

data UpdateTeamMember

See: updateTeamMember smart constructor.

Request Lenses

utmRemoteAccessAllowed :: Lens' UpdateTeamMember (Maybe Bool)

Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile. Even if this is set to True, the user must associate a public key with their profile before the user can access resources.

utmProjectRole :: Lens' UpdateTeamMember (Maybe Text)

The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

utmProjectId :: Lens' UpdateTeamMember Text

The ID of the project.

utmUserARN :: Lens' UpdateTeamMember Text

The Amazon Resource Name (ARN) of the user for whom you want to change team membership attributes.

Destructuring the Response

updateTeamMemberResponse

Arguments

:: Int
utmrsResponseStatus

-> UpdateTeamMemberResponse

Creates a value of UpdateTeamMemberResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

utmrsUserARN - The Amazon Resource Name (ARN) of the user whose team membership attributes were updated.

utmrsRemoteAccessAllowed - Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile.

utmrsProjectRole - The project role granted to the user.

utmrsResponseStatus - -- | The response status code.

data UpdateTeamMemberResponse

See: updateTeamMemberResponse smart constructor.

Response Lenses

utmrsUserARN :: Lens' UpdateTeamMemberResponse (Maybe Text)

The Amazon Resource Name (ARN) of the user whose team membership attributes were updated.

utmrsRemoteAccessAllowed :: Lens' UpdateTeamMemberResponse (Maybe Bool)

Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile.

utmrsProjectRole :: Lens' UpdateTeamMemberResponse (Maybe Text)

The project role granted to the user.

utmrsResponseStatus :: Lens' UpdateTeamMemberResponse Int

- | The response status code.