Network.AWS.CodeStar.UpdateUserProfile

Description

Updates a user's profile in AWS CodeStar. The user profile is not project-specific. Information in the user profile is displayed wherever the user's information appears to other users in AWS CodeStar.

Creating a Request

updateUserProfile

Arguments

:: Text
uupUserARN

-> UpdateUserProfile

Creates a value of UpdateUserProfile with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uupSshPublicKey - The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

uupEmailAddress - The email address that is displayed as part of the user's profile in AWS CodeStar.

uupDisplayName - The name that is displayed as the friendly name for the user in AWS CodeStar.

uupUserARN - The name that will be displayed as the friendly name for the user in AWS CodeStar.

data UpdateUserProfile

See: updateUserProfile smart constructor.

Request Lenses

uupSshPublicKey :: Lens' UpdateUserProfile (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

uupEmailAddress :: Lens' UpdateUserProfile (Maybe Text)

The email address that is displayed as part of the user's profile in AWS CodeStar.

uupDisplayName :: Lens' UpdateUserProfile (Maybe Text)

The name that is displayed as the friendly name for the user in AWS CodeStar.

uupUserARN :: Lens' UpdateUserProfile Text

The name that will be displayed as the friendly name for the user in AWS CodeStar.

Destructuring the Response

updateUserProfileResponse

Arguments

:: Int
uuprsResponseStatus

-> Text
uuprsUserARN

-> UpdateUserProfileResponse

Creates a value of UpdateUserProfileResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uuprsLastModifiedTimestamp - The date the user profile was last modified, in timestamp format.

uuprsSshPublicKey - The SSH public key associated with the user in AWS CodeStar. This is the public portion of the public/private keypair the user can use to access project resources if a project owner allows the user remote access to those resources.

uuprsEmailAddress - The email address that is displayed as part of the user's profile in AWS CodeStar.

uuprsDisplayName - The name that is displayed as the friendly name for the user in AWS CodeStar.

uuprsCreatedTimestamp - The date the user profile was created, in timestamp format.

uuprsResponseStatus - -- | The response status code.

uuprsUserARN - The Amazon Resource Name (ARN) of the user in IAM.

data UpdateUserProfileResponse

See: updateUserProfileResponse smart constructor.

Response Lenses

uuprsLastModifiedTimestamp :: Lens' UpdateUserProfileResponse (Maybe UTCTime)

The date the user profile was last modified, in timestamp format.

uuprsSshPublicKey :: Lens' UpdateUserProfileResponse (Maybe Text)

The SSH public key associated with the user in AWS CodeStar. This is the public portion of the public/private keypair the user can use to access project resources if a project owner allows the user remote access to those resources.

uuprsEmailAddress :: Lens' UpdateUserProfileResponse (Maybe Text)

The email address that is displayed as part of the user's profile in AWS CodeStar.

uuprsDisplayName :: Lens' UpdateUserProfileResponse (Maybe Text)

The name that is displayed as the friendly name for the user in AWS CodeStar.

uuprsCreatedTimestamp :: Lens' UpdateUserProfileResponse (Maybe UTCTime)

The date the user profile was created, in timestamp format.

uuprsResponseStatus :: Lens' UpdateUserProfileResponse Int

- | The response status code.

uuprsUserARN :: Lens' UpdateUserProfileResponse Text

The Amazon Resource Name (ARN) of the user in IAM.